Extract error to file
